requirements.txt
numpy
fpdf2
kaleido
plotly
pandas
streamlit
